switched to using qlib

+++ setup.py
@@ -3,10 +3,8 @@ import sys
 import os
 from os.path import relpath
 from pathlib import Path
-import huggingface_hub
 import argparse
-
-hfapi = huggingface_hub.HfApi()
+import qlib
 
 def defvar(f, name:str, value:str=None, disable=False):
     if disable or value is None:
@@ -29,42 +27,48 @@ def main():
                         help='Local affix to model name')
     parser.add_argument('--force', '-f', action='store_true',
                         help='Force overwrite existing files')
+    parser.add_argument('--description', '--desc', '-s', type=str, default=None,
+                        help='Description of the model')
+    parser.add_argument('--test', action='store_true',
+                        help='Test mode')
     args = parser.parse_args()
 
     if args.affix and not args.affix.startswith('-'):
         args.affix = '-' + args.affix
 
-
     if not '/' in (baserepo := args.basemodel.removeprefix('https://huggingface.co/')):
         raise ValueError('basemodel must be of the form owner/model')
-    author,basemodel = baserepo.split('/')[:2]
-    quantmodel = basemodel + args.affix + '-GGUF'
-    quantmodel_dir = args.build_root / quantmodel
+    
+    basemodel = qlib.Model(baserepo)
+
+    quantmodel = basemodel.model_name + args.affix
+    quantmodel_dir = args.build_root / (quantmodel + '-GGUF')
     makefile = quantmodel_dir / 'GNUmakefile'
 
     mk_dir = top / 'mk'
     defs_mk = relpath(mk_dir/'defs.mk', quantmodel_dir)
     basemodel_dir = quantmodel_dir / 'basemodel'
     basemodel_id = basemodel_dir / 'model-id.txt'
-    basemodel_link = basemodel_dir / basemodel
+    basemodel_link = basemodel_dir / basemodel.model_name
 
     basemodel_dir.mkdir(parents=True, exist_ok=True)
-    cache_path = Path(hfapi.snapshot_download(repo_id=baserepo))
+    if not args.test:
+        cache_path = basemodel.download()
 
-    if basemodel_link.is_symlink():
-        tgt = basemodel_link.readlink()
-        if tgt.samefile(cache_path):
-            cache_path = None
-        elif not tgt.exists():
-            args.force = True
+        if basemodel_link.is_symlink():
+            tgt = basemodel_link.readlink()
+            if tgt.samefile(cache_path):
+                cache_path = None
+            elif not tgt.exists():
+                args.force = True
 
-    if cache_path:
-        if basemodel_link.exists():
-            if args.force:
-                basemodel_link.unlink()
-            else:
-                raise RuntimeError(f'{basemodel_link} exists and --force not given')
-        basemodel_link.symlink_to(cache_path, target_is_directory=True)
+        if cache_path:
+            if basemodel_link.exists():
+                if args.force:
+                    basemodel_link.unlink()
+                else:
+                    raise RuntimeError(f'{basemodel_link} exists and --force not given')
+            basemodel_link.symlink_to(cache_path, target_is_directory=True)
 
     with basemodel_id.open('wt') as f:
         f.write(baserepo)
@@ -79,9 +83,11 @@ def main():
         defvar(f, 'TOASTER_ROOT', os.getenv('TOASTER_ROOT'))
         defvar(f, 'BASEREPO', baserepo)
         defvar(f, 'QUANTMODEL', quantmodel)
-        defvar(f, 'AUTHOR', author)
-        defvar(f, 'BASEMODEL', basemodel)
-        f.write(f'\n\ninclude {defs_mk}\n')
+        defvar(f, 'AUTHOR', basemodel.owner)
+        defvar(f, 'BASEMODEL', basemodel.model_name)
+        defvar(f, 'CATNAME', basemodel.catalog_name())
+        defvar(f, 'DESCRIPTION', args.description)
+        f.write(f'\ninclude {defs_mk}\n\n')
 
     print(f'Makefile in {makefile}')
